feat(initialize): initialize
