--- src/core/index.ts
@@ -1,2 +1,0 @@
-export * from "./http-status";
-export * from "./locale";
