--- src/fetch.ts
@@ -1,1 +1,0 @@
-// import "@fetchin/fetch"
