+++ src/index.ts
@@ -1,2 +1,1 @@
 export * from './fetchin'
-export * from './core'
