--- src/locale/index.ts
@@ -1,3 +1,0 @@
-export { default as enUS } from './en-us'
-export { default as zhCN } from './zh-cn'
-export { default as zhTW } from './zh-tw'
