+++ src/locales/en-us.json
@@ -1,0 +1,6 @@
+{
+  "ensign": "English",
+  "locale": {
+  },
+  "name": "en_US"
+}
