+++ src/locales/zh-cn.json
@@ -1,0 +1,6 @@
+{
+  "ensign": "中国",
+  "locale": {
+  },
+  "name": "zh_CN"
+}
