+++ src/locales/zh-tw.json
@@ -1,0 +1,6 @@
+{
+  "ensign": "中国台湾",
+  "locale": {
+  },
+  "name": "zh_TW"
+}
