+++ tsconfig.json
@@ -1,10 +1,12 @@
 {
   "extends": "@ombro/ts-config/tsconfig.esm.json",
   "compilerOptions": {
+    "baseUrl": ".",
     "composite": true,
     "removeComments": true,
+    "resolveJsonModule": true,
     "types": ["vite/client", "node", "vitest"]
   },
-  "include": ["src"],
+  "include": ["src/**/*"],
   "exclude": ["**/*.test.ts", "node_modules", "test/**"]
 }
